# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ViaX Scout SaaS API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User profile operations
  - name: analyses
    description: Route analysis operations
  - name: dashboard
    description: Dashboard and metrics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/password:
    patch:
      operationId: updatePassword
      tags: [users]
      summary: Update password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordBody"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/settings:
    get:
      operationId: getSettings
      tags: [users]
      summary: Get user settings
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
    patch:
      operationId: updateSettings
      tags: [users]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"

  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List analyses for current user
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysesList"
    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Submit a new route analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalysisBody"
      responses:
        "201":
          description: Analysis created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"

  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get a single analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAnalysis
      tags: [analyses]
      summary: Delete an analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary metrics
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/recent:
    get:
      operationId: getRecentAnalyses
      tags: [dashboard]
      summary: Get recent analyses
      responses:
        "200":
          description: Recent analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalysisSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        birthDate:
          type: ["string", "null"]
      required:
        - name
        - email
        - password

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - createdAt

    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]

    UpdatePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword

    UserSettings:
      type: object
      properties:
        parserMode:
          type: string
          enum: [builtin, ai]
          description: "builtin = embedded parser, ai = AI-powered"
        aiProvider:
          type: ["string", "null"]
          description: "AI provider name when parserMode=ai"
        aiApiKey:
          type: ["string", "null"]
          description: "API key for AI provider"
        toleranceMeters:
          type: integer
          description: "Coordinate-to-address match tolerance in meters"
      required:
        - parserMode
        - toleranceMeters

    UpdateSettingsBody:
      type: object
      properties:
        parserMode:
          type: string
          enum: [builtin, ai]
        aiProvider:
          type: ["string", "null"]
        aiApiKey:
          type: ["string", "null"]
        toleranceMeters:
          type: integer

    Analysis:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        fileName:
          type: string
        totalAddresses:
          type: integer
        nuances:
          type: integer
        geocodeSuccess:
          type: integer
        similarityAvg:
          type: number
        processingTimeMs:
          type: integer
        parserMode:
          type: string
        status:
          type: string
          enum: [pending, processing, done, error]
        results:
          type: ["string", "null"]
          description: "JSON string of detailed results"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - fileName
        - totalAddresses
        - nuances
        - geocodeSuccess
        - similarityAvg
        - processingTimeMs
        - parserMode
        - status
        - createdAt
        - updatedAt

    AnalysisSummary:
      type: object
      properties:
        id:
          type: integer
        fileName:
          type: string
        totalAddresses:
          type: integer
        nuances:
          type: integer
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - fileName
        - totalAddresses
        - nuances
        - status
        - createdAt

    AnalysesList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Analysis"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - page
        - limit

    CreateAnalysisBody:
      type: object
      properties:
        fileName:
          type: string
        totalAddresses:
          type: integer
        nuances:
          type: integer
        geocodeSuccess:
          type: integer
        similarityAvg:
          type: number
        processingTimeMs:
          type: integer
        parserMode:
          type: string
        results:
          type: ["string", "null"]
      required:
        - fileName
        - totalAddresses
        - nuances
        - geocodeSuccess
        - similarityAvg
        - processingTimeMs
        - parserMode

    DashboardSummary:
      type: object
      properties:
        totalAnalyses:
          type: integer
        totalAddressesProcessed:
          type: integer
        avgNuanceRate:
          type: number
        avgGeocodeSuccess:
          type: number
        avgSimilarity:
          type: number
        analysesThisMonth:
          type: integer
      required:
        - totalAnalyses
        - totalAddressesProcessed
        - avgNuanceRate
        - avgGeocodeSuccess
        - avgSimilarity
        - analysesThisMonth