# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ViaX Scout SaaS API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const RegisterBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
  birthDate: zod.string().nullish(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    avatarUrl: zod.string().nullish(),
    birthDate: zod.string().nullish(),
    createdAt: zod.string(),
  }),
  message: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  avatarUrl: zod.string().nullish(),
  birthDate: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Update user profile
 */
export const UpdateProfileBody = zod.object({
  name: zod.string().optional(),
  avatarUrl: zod.string().nullish(),
  birthDate: zod.string().nullish(),
});

export const UpdateProfileResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  avatarUrl: zod.string().nullish(),
  birthDate: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Update password
 */
export const UpdatePasswordBody = zod.object({
  currentPassword: zod.string(),
  newPassword: zod.string(),
});

export const UpdatePasswordResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get user settings
 */
export const GetSettingsResponse = zod.object({
  parserMode: zod
    .enum(["builtin", "ai"])
    .describe("builtin = embedded parser, ai = AI-powered"),
  aiProvider: zod
    .string()
    .nullish()
    .describe("AI provider name when parserMode=ai"),
  aiApiKey: zod.string().nullish().describe("API key for AI provider"),
  toleranceMeters: zod
    .number()
    .describe("Coordinate-to-address match tolerance in meters"),
});

/**
 * @summary Update user settings
 */
export const UpdateSettingsBody = zod.object({
  parserMode: zod.enum(["builtin", "ai"]).optional(),
  aiProvider: zod.string().nullish(),
  aiApiKey: zod.string().nullish(),
  toleranceMeters: zod.number().optional(),
});

export const UpdateSettingsResponse = zod.object({
  parserMode: zod
    .enum(["builtin", "ai"])
    .describe("builtin = embedded parser, ai = AI-powered"),
  aiProvider: zod
    .string()
    .nullish()
    .describe("AI provider name when parserMode=ai"),
  aiApiKey: zod.string().nullish().describe("API key for AI provider"),
  toleranceMeters: zod
    .number()
    .describe("Coordinate-to-address match tolerance in meters"),
});

/**
 * @summary List analyses for current user
 */
export const listAnalysesQueryPageDefault = 1;
export const listAnalysesQueryLimitDefault = 10;

export const ListAnalysesQueryParams = zod.object({
  page: zod.coerce.number().default(listAnalysesQueryPageDefault),
  limit: zod.coerce.number().default(listAnalysesQueryLimitDefault),
});

export const ListAnalysesResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.number(),
      fileName: zod.string(),
      totalAddresses: zod.number(),
      nuances: zod.number(),
      geocodeSuccess: zod.number(),
      similarityAvg: zod.number(),
      processingTimeMs: zod.number(),
      parserMode: zod.string(),
      status: zod.enum(["pending", "processing", "done", "error"]),
      results: zod
        .string()
        .nullish()
        .describe("JSON string of detailed results"),
      createdAt: zod.string(),
      updatedAt: zod.string(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
});

/**
 * @summary Submit a new route analysis
 */
export const CreateAnalysisBody = zod.object({
  fileName: zod.string(),
  totalAddresses: zod.number(),
  nuances: zod.number(),
  geocodeSuccess: zod.number(),
  similarityAvg: zod.number(),
  processingTimeMs: zod.number(),
  parserMode: zod.string(),
  results: zod.string().nullish(),
});

/**
 * @summary Get a single analysis
 */
export const GetAnalysisParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnalysisResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  fileName: zod.string(),
  totalAddresses: zod.number(),
  nuances: zod.number(),
  geocodeSuccess: zod.number(),
  similarityAvg: zod.number(),
  processingTimeMs: zod.number(),
  parserMode: zod.string(),
  status: zod.enum(["pending", "processing", "done", "error"]),
  results: zod.string().nullish().describe("JSON string of detailed results"),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete an analysis
 */
export const DeleteAnalysisParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get dashboard summary metrics
 */
export const GetDashboardSummaryResponse = zod.object({
  totalAnalyses: zod.number(),
  totalAddressesProcessed: zod.number(),
  avgNuanceRate: zod.number(),
  avgGeocodeSuccess: zod.number(),
  avgSimilarity: zod.number(),
  analysesThisMonth: zod.number(),
});

/**
 * @summary Get recent analyses
 */
export const GetRecentAnalysesResponseItem = zod.object({
  id: zod.number(),
  fileName: zod.string(),
  totalAddresses: zod.number(),
  nuances: zod.number(),
  status: zod.string(),
  createdAt: zod.string(),
});
export const GetRecentAnalysesResponse = zod.array(
  GetRecentAnalysesResponseItem,
);